% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtool.R
\name{addGeneName4Ensembl}
\alias{addGeneName4Ensembl}
\title{Add gene nemes}
\usage{
addGeneName4Ensembl(mart, report = "report")
}
\arguments{
\item{mart}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{report}{The report directory for adding gene names.}
}
\value{
none
}
\description{
Add "gene name" column to the attachments (eg. snv.tsv and idl.tsv) of report 
if there is a demand (optional function). This function is useful for "Ensembl" IDs.
}
\examples{
\dontrun{
library(PGA)
# set the biomaRt parameters just as "PrepareAnnotationEnsembl2" did
 
mart <- biomaRt::useMart("ENSEMBL_MART_ENSEMBL",
        dataset="hsapiens_gene_ensembl",
        host="grch37.ensembl.org", 
        path="/biomart/martservice",
        archive=FALSE)
addGeneName4Ensembl(mart=mart,report="report")

}
}
