% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreator.R
\name{buildFusionProteinDB}
\alias{buildFusionProteinDB}
\title{Create customized protein database from fusion events}
\usage{
buildFusionProteinDB(x, species = "Homo sapiens",
  genome_version = "hg38", fusion_method = "STAR-Fusion",
  max_nt = 100, out_dir = "./", prefix = "fusion",
  translating_method = "six_frame", min_aa_length = 10)
}
\arguments{
\item{x}{A tsv format file which contains fusion events.}

\item{species}{Species, default is "Homo sapiens"}

\item{genome_version}{Genome version, default is "hg38"}

\item{fusion_method}{Fusion calling method, default is "STAR-Fusion"}

\item{max_nt}{The max length of DNA sequences to be extracted for each side,
default is 60}

\item{out_dir}{Output directory}

\item{prefix}{The prefix of output files}

\item{translating_method}{Translating DNA to protein (six_frame,three_frame,
longest), default is six_frame.}

\item{min_aa_length}{The minimum length of proteins, default is 10 aa.}
}
\value{
The database file
}
\description{
Create customized protein database from fusion events
}
\examples{
fusion_file <- system.file("extdata/fusion/", "star-fusion_example_input.tsv",package="PGA")
# This example input was downloaded from STAR-Fusion website (https://github.com/STAR-Fusion/STAR-Fusion/wiki)
buildFusionProteinDB(fusion_file,genome_version="hg19")
}
