% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsearch.R
\name{calculateFDR}
\alias{calculateFDR}
\title{Perform global or separate FDR estimation}
\usage{
calculateFDR(psmfile = NULL, db = NULL, fdr = 0.01,
  peptide_level = FALSE, decoyPrefix = "###REV###",
  novelPrefix = NULL, better_score_lower = TRUE, remap = FALSE,
  out_dir = "./", protein_inference = FALSE, score_t = 1, xmx = 2,
  verbose = 1)
}
\arguments{
\item{psmfile}{PSM file in TSV format}

\item{db}{A FASTA format database file used for MS/MS searching.}

\item{fdr}{FDR cutoff, default is 0.01}

\item{peptide_level}{Peptide level FDR, default is FALSE}

\item{decoyPrefix}{The prefix of decoy sequences ID. Default is "###REV###".
"###REV###" is the prefix which used by function \code{dbCreator}.}

\item{novelPrefix}{The prefix of novel protein ID. Default is NULL. If the 
value is NULL, it will perform global FDR estimation. Otherwise, it will 
perform separate FDR estimation for novel peptides.}

\item{better_score_lower}{TRUE: lower score is better, FALSE: higher score is better.
Default is TRUE.}

\item{remap}{TRUE: re-map peptide to protein, 
FALSE: use the peptide protein mapping data in the PSM file. Default is FALSE.}

\item{out_dir}{Output directory.}

\item{protein_inference}{Whether or not to perform protein inference. Default is 
FALSE}

\item{score_t}{Score transformation for score distribution plot. 
0: no transformation, 1: -log(score).}

\item{xmx}{The maximum Java heap size. The unit is "G". Default is 2.}

\item{verbose}{Output level. Default is 1.}
}
\value{
none
}
\description{
Perform global or separate FDR estimation
}
