% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtool.R
\name{createProDB4DenovoRNASeq}
\alias{createProDB4DenovoRNASeq}
\title{Create protein database based on the transcripts from de novo 
reconstruction of transcriptome from RNA-seq data}
\usage{
createProDB4DenovoRNASeq(infa = "./trinity.fasta",
  bool_use_3frame = FALSE, outmtab = "novel_transcripts_ntx.tab",
  outfa = "./novel_transcripts_ntx.fasta", bool_get_longest = TRUE,
  make_decoy = TRUE, decoy_tag = "#REV#", outfile_name = "test")
}
\arguments{
\item{infa}{A FASTA format file containing transcript sequences 
(such as derived from de novo transcriptome assembly by Trinity)}

\item{bool_use_3frame}{A logical variable indicating whether to translate the 
raw sequences with 3-frame (forward). Default is 6-frame translation (FALSE).}

\item{outmtab}{A txt format file containing the novel transcripts information}

\item{outfa}{The output fasta format protein sequence file}

\item{bool_get_longest}{When it's set as TRUE, only the longest protein sequences will be 
retained after the raw sequences are translated with 3-frame or 6-frame. 
Otherwise, all the protein sequences with longer than 30 aa will be retained.}

\item{make_decoy}{A logical variable indicating whether to add the decoy sequences.}

\item{decoy_tag}{The prefix of decoy sequence IDs.}

\item{outfile_name}{Output file name}
}
\value{
The database file(s)
}
\description{
Create protein database based on the transcripts from de novo 
reconstruction of transcriptome from RNA-seq data.
}
\examples{
transcript_seq_file <- system.file("extdata/input", "Trinity.fasta",package="PGA")
createProDB4DenovoRNASeq(infa=transcript_seq_file)
}
