% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbCreator.R
\name{dbCreator}
\alias{dbCreator}
\title{Create customized protein database from RNA-Seq data}
\usage{
dbCreator(gtfFile = NULL, vcfFile = NULL, bedFile = NULL,
  tabFile = NULL, annotation_path = NULL, outdir, outfile_name,
  lablersid = FALSE, COSMIC = FALSE, bool_get_longest = TRUE,
  organism = "Homo sapiens", make_decoy = TRUE, genome = NULL,
  debug = FALSE, ...)
}
\arguments{
\item{gtfFile}{A GTF format file containing novel transcripts information}

\item{vcfFile}{A VCF format file containing SNV and INDEL information}

\item{bedFile}{A BED format file from Tophat2 containing juction information.}

\item{tabFile}{A Tab format file from Hisat2 containing juction information (e.g. splicesites.tab). As HISAT2 is a successor to both HISAT and TopHat2, we recommend that users switch to this option.}

\item{annotation_path}{This directory contains numerous pieces of genome 
annotation information which can be downloaded by 
\code{\link{PrepareAnnotationEnsembl2}} or \code{\link{PrepareAnnotationRefseq2}}.}

\item{outdir}{Output directory}

\item{outfile_name}{Output file name}

\item{lablersid}{A logical indicating whether to do the SNV annotation(dbSNP)}

\item{COSMIC}{A logical indicating whether to do the SNV annotation(COSMIC)}

\item{bool_get_longest}{When it's set as TRUE, the longest sequences will be 
retained after the DNA sequences are six-frame translated into protein 
sequences. Otherwise, the protein sequences more than 30 aa are retained.}

\item{organism}{What is the Genus and species of this organism.Please use 
proper scientific nomenclature for example: "Homo sapiens" and not "human",
default is "Homo sapiens".}

\item{make_decoy}{A logical indicating whether to add the decoy sequences}

\item{genome}{Genome information. This is a BSgenome object(e.g. Hsapiens). 
Default is NULL.}

\item{debug}{Output the debug info (Default is <False>).}

\item{...}{Additional arguments}
}
\value{
The database file
}
\description{
The main function to create customized protein database from 
RNA-Seq data
}
\examples{
vcffile <- system.file("extdata/input", "PGA.vcf",package="PGA")
bedfile <- system.file("extdata/input", "junctions.bed",package="PGA")
gtffile <- system.file("extdata/input", "transcripts.gtf",package="PGA")
annotation <- system.file("extdata", "annotation",package="PGA")
outfile_path<-"db/"
outfile_name<-"test"
library(BSgenome.Hsapiens.UCSC.hg19)
dbCreator(gtfFile=gtffile,vcfFile=vcffile,bedFile=bedfile,
          annotation_path=annotation,outfile_name=outfile_name,
          genome=Hsapiens,outdir=outfile_path)
}
