% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRun.R
\name{easyRun}
\alias{easyRun}
\title{easyRun}
\usage{
easyRun(gtfFile = NULL, vcfFile = NULL, bedFile = NULL,
  spectra = NULL, annotation_path = NULL, outdir = "pga_dir",
  outPrefix = "pga", lablersid = FALSE, COSMIC = FALSE,
  bool_get_longest = TRUE, organism = "Homo sapiens", genome = NULL,
  enzyme = "[KR]|[X]", tol = 10, tolu = "ppm", itol = 0.6,
  itolu = "Daltons", varmod = NULL, fixmod = NULL, miss = 2,
  maxCharge = 8, ti = FALSE, cpu = 0, fdr = 0.01, alignment = 1,
  xmx = 2, ...)
}
\arguments{
\item{gtfFile}{A GTF format file containing novel transcripts information}

\item{vcfFile}{A VCF format file containing SNV and INDEL information}

\item{bedFile}{A BED format file containing juction information}

\item{spectra}{MS/MS peak list file}

\item{annotation_path}{This directory contains numerous pieces of genome 
annotation information which can be downloaded by 
\code{\link{PrepareAnnotationEnsembl2}} or \code{\link{PrepareAnnotationRefseq2}}.}

\item{outdir}{Output directory.}

\item{outPrefix}{The prefix of output file.}

\item{lablersid}{A logical indicating whether to do the SNV annotation(dbSNP)}

\item{COSMIC}{A logical indicating whether to do the SNV annotation(COSMIC)}

\item{bool_get_longest}{When it's set as TRUE, the longest sequences will be 
retained after the DNA sequences are six-frame translated into protein 
sequences. Otherwise, the protein sequences more than 30 aa are retained.}

\item{organism}{What is the Genus and species of this organism.Please use 
proper scientific nomenclature for example: "Homo sapiens" and not "human",
default is "Homo sapiens".}

\item{genome}{Genome information. This is a BSgenome object(e.g. Hsapiens).}

\item{enzyme}{Specification of specific protein cleavage sites. 
Default is "[KR]|[X]".}

\item{tol}{Parent ion mass tolerance (monoisotopic mass).}

\item{tolu}{Parent ion M+H mass tolerance window units.}

\item{itol}{Fragment ion mass tolerance (monoisotopic mass).}

\item{itolu}{Unit for fragment ion mass tolerance (monoisotopic mass).}

\item{varmod}{Specificiation of potential modifications of residues.}

\item{fixmod}{Specification of modifications of residues.}

\item{miss}{The number of missed cleavage sites. Default is 2.}

\item{maxCharge}{The Maximum parent charge, default is 8}

\item{ti}{anticipate carbon isotope parent ion assignment errors.
Default is false.}

\item{cpu}{The number of CPU used for X!Tandem search. Default is 1.}

\item{fdr}{FDR for peptide identification. Default is 0.01 at PSM level.}

\item{alignment}{0 or 1 to determine if peptide should be alignment or not.
Default is 0.}

\item{xmx}{The maximum Java heap size. The unit is "G".}

\item{...}{Additional arguments}
}
\value{
none
}
\description{
This function is used to automate the peptide identification 
based on searching the customized database derived from RNA-Seq data.
}
\examples{
vcffile <- system.file("extdata/input", "PGA.vcf",package="PGA")
bedfile <- system.file("extdata/input", "junctions.bed",package="PGA")
gtffile <- system.file("extdata/input", "transcripts.gtf",package="PGA")
annotation <- system.file("extdata", "annotation",package="PGA")
library(BSgenome.Hsapiens.UCSC.hg19)
msfile <- system.file("extdata/input", "pga.mgf",package="PGA")
easyRun(gtfFile=gtffile,vcfFile=vcffile,bedFile=bedfile,spectra=msfile,
        annotation_path=annotation,genome=Hsapiens,cpu = 6,
        enzyme = "[KR]|[X]", varmod = "15.994915@M",itol = 0.05,
        fixmod = "57.021464@C", tol = 10, tolu = "ppm", itolu = "Daltons", 
        miss = 2, maxCharge = 8, ti = FALSE,xmx=1)
}
