\name{get.topfew}
\alias{get.topfew}
\title{Find the Top Few Ranked Variables}

\description{
Find the top few variables ranked by \code{\link{pga}}. 
}

\usage{
get.topfew(obj, num)
}

\arguments{
            
\item{obj}{object created by \code{pga}.}

\item{num}{number of top variables to retrieve.}

}

\seealso{
\code{\link{pga}},
\code{\link{view.bubble}},
\code{\link{get.pga.ans}}
}

\author{Mu Zhu, University of Waterloo, Canada.}

\keyword{misc}
