% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtool.R
\name{PrepareAnnotationEnsembl2}
\alias{PrepareAnnotationEnsembl2}
\title{Prepare annotation from ENSEMBL}
\usage{
PrepareAnnotationEnsembl2(mart, annotation_path, splice_matrix = FALSE,
  dbsnp = NULL, transcript_ids = NULL, COSMIC = FALSE, ...)
}
\arguments{
\item{mart}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{annotation_path}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{splice_matrix}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{dbsnp}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{transcript_ids}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{COSMIC}{See detail in function \code{\link{PrepareAnnotationEnsembl}}.}

\item{...}{Additional arguments}
}
\value{
Several .RData file containing annotations needed for following analysis.
}
\description{
Prepare the annotation from ENSEMBL through biomaRt. This 
function is modified from the function \code{\link{PrepareAnnotationEnsembl}} in \pkg{customProDB}.
}
\examples{
ensembl <- biomaRt::useMart("ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl",
                    host="grch37.ensembl.org", path="/biomart/martservice",
                    archive=FALSE)

annotation_path <- tempdir()
transcript_ids <- c("ENST00000234420", "ENST00000269305", "ENST00000445888")

PrepareAnnotationEnsembl2(mart=ensembl, annotation_path=annotation_path,
                          splice_matrix=FALSE, dbsnp=NULL, transcript_ids=transcript_ids,
                          COSMIC=FALSE)
}
\seealso{
\code{\link{PrepareAnnotationRefseq2}}.
}

