% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtool.R
\name{PrepareAnnotationRefseq2}
\alias{PrepareAnnotationRefseq2}
\title{Prepare annotation from Refseq}
\usage{
PrepareAnnotationRefseq2(genome = "hg19", CDSfasta, pepfasta, annotation_path,
  dbsnp = NULL, transcript_ids = NULL, splice_matrix = FALSE,
  COSMIC = FALSE, ...)
}
\arguments{
\item{genome}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{CDSfasta}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{pepfasta}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{annotation_path}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{dbsnp}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{transcript_ids}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{splice_matrix}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{COSMIC}{See detail in function \code{\link{PrepareAnnotationRefseq}}.}

\item{...}{Additional arguments}
}
\value{
Several .RData file containing annotations needed for following analysis.
}
\description{
Prepare the annotation for Refseq through UCSC table browser. This 
function is modified from the function \code{\link{PrepareAnnotationRefseq}} in \pkg{customProDB}.
}
\examples{
\dontrun{
transcript_ids <- c("NM_001126112", "NM_033360", "NR_073499")
pepfasta <- system.file("extdata", "refseq_pro_seq.fasta",
                        package="customProDB")
CDSfasta <- system.file("extdata", "refseq_coding_seq.fasta",
                        package="customProDB")
annotation_path <- tempdir()
PrepareAnnotationRefseq2(genome='hg19', CDSfasta, pepfasta, annotation_path,
                        dbsnp=NULL, transcript_ids=transcript_ids,
                        splice_matrix=FALSE, COSMIC=FALSE)
}
}
\seealso{
\code{\link{PrepareAnnotationEnsembl2}}.
}
