% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportGear}
\alias{reportGear}
\title{The main function for report generation}
\usage{
reportGear(parser_dir, tab_dir, report_dir)
}
\arguments{
\item{parser_dir}{The directory which contains the peptide identification 
results}

\item{tab_dir}{The directory which contains the database annotation files}

\item{report_dir}{The report output directory}
}
\value{
none
}
\description{
The main function for report generation
}
\examples{
vcffile <- system.file("extdata/input", "PGA.vcf",package="PGA")
bedfile <- system.file("extdata/input", "junctions.bed",package="PGA")
gtffile <- system.file("extdata/input", "transcripts.gtf",package="PGA")
annotation <- system.file("extdata", "annotation",package="PGA")
outfile_path<-"db/"
outfile_name<-"test"
library(BSgenome.Hsapiens.UCSC.hg19)
dbfile <- dbCreator(gtfFile=gtffile,vcfFile=vcffile,bedFile=bedfile,
                    annotation_path=annotation,outfile_name=outfile_name,
                    genome=Hsapiens,outdir=outfile_path)
          
msfile <- system.file("extdata/input", "pga.mgf",package="PGA")
idfile <- runTandem(spectra = msfile, fasta = dbfile, outdir = "./", cpu = 6,
                    enzyme = "[KR]|[X]", varmod = "15.994915@M",itol = 0.05,
                    fixmod = "57.021464@C", tol = 10, tolu = "ppm", 
                    itolu = "Daltons", miss = 2, maxCharge = 8, ti = FALSE)
parserGear(file = idfile, db = dbfile, decoyPrefix="#REV#",xmx=1,thread=8,
           outdir = "parser_outdir")
reportGear(parser_dir = "parser_outdir", tab_dir = outfile_path,
           report_dir = "report")
}
