% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsearch.R
\name{runTandem}
\alias{runTandem}
\title{run X!Tandem}
\usage{
runTandem(spectra = "", fasta = "", outdir = ".", cpu = 1,
  enzyme = "[KR]|[X]", tol = 10, tolu = "ppm", itol = 0.6,
  itolu = "Daltons", varmod = NULL, fixmod = NULL, miss = 2,
  maxCharge = 8, ti = FALSE)
}
\arguments{
\item{spectra}{MS/MS peak list file}

\item{fasta}{Protein database file for searching.}

\item{outdir}{The output directory.}

\item{cpu}{The number of CPU used for X!Tandem search. Default is 1.}

\item{enzyme}{Specification of specific protein cleavage sites.
Default is "[KR]|[X]".}

\item{tol}{Parent ion mass tolerance (monoisotopic mass).}

\item{tolu}{Parent ion M+H mass tolerance window units.}

\item{itol}{Fragment ion mass tolerance (monoisotopic mass).}

\item{itolu}{Unit for fragment ion mass tolerance (monoisotopic mass).}

\item{varmod}{Specificiation of potential modifications of residues.}

\item{fixmod}{Specification of modifications of residues.}

\item{miss}{The number of missed cleavage sites. Default is 2.}

\item{maxCharge}{The Maximum parent charge, default is 8}

\item{ti}{anticipate carbon isotope parent ion assignment errors.
Default is false.}
}
\value{
The search result file path
}
\description{
run X!Tandem
}
\examples{
vcffile <- system.file("extdata/input", "PGA.vcf",package="PGA")
bedfile <- system.file("extdata/input", "junctions.bed",package="PGA")
gtffile <- system.file("extdata/input", "transcripts.gtf",package="PGA")
annotation <- system.file("extdata", "annotation",package="PGA")
outfile_path<-"db/"
outfile_name<-"test"
library(BSgenome.Hsapiens.UCSC.hg19)
dbfile <- dbCreator(gtfFile=gtffile,vcfFile=vcffile,bedFile=bedfile,
                    annotation_path=annotation,outfile_name=outfile_name,
                    genome=Hsapiens,outdir=outfile_path)

msfile <- system.file("extdata/input", "pga.mgf",package="PGA")
runTandem(spectra = msfile, fasta = dbfile, outdir = "./", cpu = 6,
          enzyme = "[KR]|[X]", varmod = "15.994915@M",
          fixmod = "57.021464@C", tol = 10, tolu = "ppm", itol = 0.05,
          itolu = "Daltons", miss = 2, maxCharge = 8, ti = FALSE)
}

