\name{summarize.PGAsim}
\alias{summarize.PGAsim}
\title{Utility Function for Summarizing Simulation Results}

\description{
This function is used to produce \cite{Technometrics}, 48, page 499, Table 
9. Please see examples at \link{pga-package} for how to use this function.
}

\seealso{
\code{\link{get.pga.ans}},
\code{\link{get.topfew}}
}

\author{Mu Zhu, University of Waterloo, Canada.}

\keyword{misc}    

